Changes to instance image whacking...
Modify instance images to have 3 NICs dhcp instead of one/two.

The default Ubuntu cloud images only support 1 NIC, by default (meaning
that only /etc/network/interfaces.d/eth1.cfg is set up).

Also boost the MTU of my static images in aarch64.

Also, fix the root password in the x86 image to be the one in the
profile instructions.

Also change the name of the "best" aarch64 to "trusty-server" so it
matches the x8664 profile.