Serial console support.
Openstack doesn't yet have 1) serial console client builtin to
dashboard, nor 2) serial console log + r/w serial console access.

So right now, you have to choose if you want to enable r/w console
access via CLI client, or if you want to be able to view serial logs in
the dashboard web UI; default is the latter (logs).

If you enable r/w consoles, we download a simple websocket console
client, setup a little frontend script, so users can type

  $ /root/setup/novaconsole.sh <instancename>

and get to the console.  Escape is ~.  In addition to the console CLI
tool, we have to grab the latest version of the python websocket
library, cause the one in Ubuntu 15.04 is horribly out of date and
doesn't seem to support binary/base64 websockets, which the console CLI
tool requires (as does the server).