Miscellaneous ceilometer improvements.
Switch to using the newer messagingv2 notification_driver (I note
they've updated the Juno docs to use this too, so maybe it was available
then and the docs just weren't updated).

Handle a few Kilo-new config file things.  They didn't seem to be
hurting anything to do it the Juno way, but now it's "correct".

Make sure to add the glance registry, not just the api, for images.