Try really hard to start rabbitmq up correctly.
Mostly, we keep trying to start the app until it succeeds.  It seems
that the server needs some time to get started before it's safe to
start up the app.  Could be as long as 5-6 seconds.  We (and openstack)
can't do anything if the app isn't running in the server.