Make package installation optional, and separate install/upgrade.
Quit trying to apt-get packages if they're installed, unless the
user selects the new DO_APT_UPGRADE option.  Always install was nice
in the beginning, but it is no longer the best use case, and it can
cause uncertainty when failures happen (i.e., if new versions of
packages get installed that the scripts can't handle).  So now there
are three apt options in the scripts and in the geni-lib script:

DO_APT_UPDATE -- updates the apt cache (often hard to do pkg
  install/upgrade if the cache is out of date); defaults to 1
DO_APT_INSTALL -- if this is set 0, we don't install *anything*
  other than critical deps (think python-m2crypto); defaults to 1
DO_APT_UPGRADE -- if this is set 1, we always run apt-get install
  to either install and/or upgrade OpenStack packages and deps.
  The big change is that this now defaults to 0 -- so packages are
  not upgraded from their current versions if they exist.