Disable memcache for Keystone by default.
Memcache + Keystone + WSGI/Apache seems to cause a problem where
Keystone is effectively unavailable (internal errors) for about a
minute... then it comes back by itself.

So we disable it by default.  The docs default to using it, but this is
far from the first time the doc defaults trigger bugs or are simply bad
configuration!