Patch oslo.service to avoid nasty hangs on service stop.
Apparently oslo.service handles stop signals (like those from systemd)
in a bad way.  When systemd stops a service, it can drive the service
into an exception (in one of its threads, I think) and I assume systemd
waits until its timeout and then kills it more firmly.  This patch went
in to oslo.service 1.2.0 or so, but of course we're stuck back in the
stone age still, around 0.9 or whatever.  So apply the patch, stupid.
We start and stop services a lot as we're setting up.