On Liberty aarch64, enable vnc using vga driver disabling usb input.
Liberty didn't seem to like the disable_vnc flag to Nova on aarch64 that
we relied on so that images would boot.  Fortuitously, qemu/libvirt have
been upgraded enough so that you can actually attach a VGA adapter to an
aarch64 KVM qemu instance.  So we do that, and now we mark images with a
specific flag that says to use the vga display driver instead of the
'cirrus' default, which qemu/libvirt aarch64 does *not* support.

Probably I should just find a way to fix the vnc disablement :).