Fix bug from the split ctl/nm days; also parse routable_pools "safely".
This was a holdover from the original profile with split ctl and nm
management nodes.  Now there is just a single ctl node by default, but
the IPs are still requested tied to the nm node.  So we were ignoring
the split and the custom naming options.

Then, the hunk o' perl that was pulling the addrs out of the manifest
was first just assuming the first cluster manifest had the proper addrs,
which is wrong; we have to look at all cluster manifests.  Then, we have
to ensure we are extracting public addrs for the NM node (which in the
non-split world is the ctl node).

So now the hunk o' perl is really large and is "safer".