Fix new "requirement" triggered by neutron-ovs-cleanup.
Some services assume they can resolve the hostname prior to network
being up (i.e. neutron-ovs-cleanup).  neutron-ovs-cleanup in particular
seems to not allow the ovs bridges to forward until it finishes.  This
is very bad.  So, to workaround, put the real control net ip and FQDN
into /etc/hosts too, both statically at setup (so it is there for the
first reboot), and in the rc.hostnames hook (so it is there after future
reboots).