Designate fixes and extras.
First: the Pike docs are quite wrong; they are written assuming presence
of a keystone v2 endpoint, which hasn't been present for several
releases (in the standard docs, that is).  Second, all the Designate
servers/clients seem to require the Designate endpoint to be /v2, not
just /.

After that, there's a bunch of Neutron/Nova integration (i.e. so that
instances can resolve names), and a few minor fixes.