Be much more careful about using Designate DNS in /etc/resolv.conf .
First, we couldn't do the modification where we were doing it; too many
configuration things could still be in flight.  But we have to ensure
that Designate is properly setup to forward to our real DNS server; and
that we succesfully found our real DNS server.  Finally, we cannot use
/var/emulab/boot/bossip; what if we have a subboss.