Add port_security_enabled=False default patch(es) for manila client.
Again, we do this because the manila client does not try to default
per-port port_security_enabled to whatever the network value for that
setting is.  Other neutron clients take the default value from the
network, or at least don't allow the network value to be overridden with
enabled if the network value is disabled.

Also, I forgot this patch for Rocky.