Add python3 support to our various scripts.
(Note that if the system default python is python2, the call to getmanifests.py
will use python2, so test-common.py must continue to support python2.  But
when setup-user-info.py goes to add ssh keys via the novaclient library, for
release >= Stein, that is python3.  Thus test-common.py must also be python3-
compat.)